Refactor package structure: move all classes to edu.ntnu.idi.idatt.boardgame
