+++ pom.xml
@@ -3,11 +3,11 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>edu.idi.idatt</groupId>
-  <artifactId>idatt2003-mappe-2025-gr57</artifactId>
+  <groupId>edu.ntnu.idi.idatt</groupId>
+  <artifactId>boardgame</artifactId>
   <version>1.0-SNAPSHOT</version>
 
-  <name>idatt2003-mappe-2025-gr57</name>
+  <name>boardgame</name>
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -100,7 +100,7 @@
                 </execution>
             </executions>
           <configuration>
-            <mainClass>edu.idi.idatt.App</mainClass>
+            <mainClass>edu.ntnu.idi.idatt.boardgame.App</mainClass>
           </configuration>
         </plugin>
       </plugins>
