+++ src/main/java/edu/ntnu/idi/idatt/boardgame/model/GameBoard.java
@@ -1,4 +1,4 @@
-package edu.idi.idatt.model;
+package edu.ntnu.idi.idatt.boardgame.model;
 
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.Pane;
